requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pyodbc
azure-storage-blob
openai
